% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_headers_dev}
\alias{auth_headers_dev}
\title{Initiate root URL with API key (Development)}
\usage{
auth_headers_dev(
  api_key = get_key(),
  NATURECUBE_URL = Sys.getenv("NATURECUBE_SIT_URL", unset =
    "https://sit.naturecube.io/api/")
)
}
\arguments{
\item{api_key}{A valid API key. If omitted, reads \code{NATURECUBE_API_KEY}
via \code{get_key()}.}

\item{NATURECUBE_URL}{The base URL for the SIT API. If omitted, reads
`NATURECUBE_SIT_URL` and falls back to the SIT default URL.}
}
\value{
A list containing the root URL and the API key
}
\description{
Creates a base URL object for the development NatureCube API.
Requires a project API key.
}
\examples{
\dontrun{
  headers <- auth_headers_dev("your_api_key")
}

}
\author{
Adam Varley
}
