% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_endpoints.R
\name{getIUCNLabels}
\alias{getIUCNLabels}
\title{Get labels from the wider IUCN database (all species)}
\usage{
getIUCNLabels(hdr, offset, search_term = NULL)
}
\arguments{
\item{hdr}{A base URL provided and valid API key returned by the function \link{auth_headers}}

\item{offset}{An integer specifying the offset for the query}

\item{search_term}{A character vector specifying the search term to be used (optional)}
}
\value{
A list containing tabular data and pagination information for iterative calls
}
\description{
Retrieve labels from the wider IUCN database, optionally filtered by a search term.
}
\examples{
\dontrun{
  getIUCNLabels(headers, offset=0, search_term="horse")
}

}
\author{
Adam Varley
}
