% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_endpoints.R
\name{get_media_assets}
\alias{get_media_assets}
\title{Retrieve media assets for a given project system record ID}
\usage{
get_media_assets(hdr, datatype = c("video", "audio", "image", "eDNA"), psrID)
}
\arguments{
\item{hdr}{A base URL provided and valid API key returned by the
function \link{auth_headers}}

\item{datatype}{A character vector of data types
c("video","audio","image","eDNA")}

\item{psrID}{A numeric vector of project system record IDs (as found in
\code{stations$project_system_record_id}).}
}
\value{
A tibble of media assets for the specified project system record
}
\description{
Get all of the station data associated with your project.
For data types c("video","audio","image")
}
\examples{
\dontrun{
  assets <- get_media_assets(headers, datatype = "video", psrID = stations$project_system_record_id)
}

}
\author{
Adam Varley
}
