% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okalaR-package.R
\docType{package}
\name{NatureCubeR-package}
\alias{NatureCubeR}
\alias{NatureCubeR-package}
\title{NatureCubeR: API wrapper around the NatureCube dashboard interface}
\description{
Allows basic reading and writing of data to and from your project on the NatureCube dashboard.
}
\author{
\strong{Maintainer}: Adam Varley \email{adam@okala.io}

Authors:
\itemize{
  \item Adam Varley \email{adam@okala.io}
}

}
\keyword{internal}
